make it a python package
