+++ setup.py
@@ -40,10 +40,17 @@ setup(
     install_requires=pkgs,
     dependency_links=new_links,
     version='0.1.0',
-    description='A awesome word alignment tool',
+    license='BSD 3-Clause',
+    description='An awesome word alignment tool',
     packages=find_packages(
         include = ['awesome_align', 'awesome_align.*'],
         exclude = ['run_align.py', 'run_train.py']
     ),
+    entry_points={
+        "console_scripts": [
+            "awesome-align=awesome_align.run_align:main",
+            "awesome-train=awesome_align.run_train:main",
+        ],
+    },
     url='https://github.com/neulab/awesome-align',
 )
